requirements.txt
Flask==3.1.0
flask-cors==5.0.1
google-cloud-vision==3.10.1
google-cloud-storage==3.1.0
gunicorn==21.2.0
pillow==11.1.0
pillow-heif==0.22.0
python-dotenv==1.1.0
requests==2.32.3
Werkzeug==3.1.0
Jinja2>=3.1.2
MarkupSafe>=2.0.1 